requirements.txt
asgiref==3.7.2
autopep8==2.0.3
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pycodestyle==2.11.0
python-dotenv==1.0.0
